+++ src/presentation/protocols/index.ts
@@ -1,3 +1,4 @@
 /* istanbul ignore file */
 export * from './controller'
 export * from './https'
+export * from './validation'
